project-root/requirements.txt
flask
python-dotenv
google-generativeai
twilio